+++ joblib/_store_backends.py
@@ -474,9 +474,14 @@ class FileSystemStoreBackend(StoreBackendBase, StoreBackendMixin):
             if os.path.dirname(location) and os.path.basename(location) == "joblib"
             else location
         )
-        with open(os.path.join(cache_directory, ".gitignore"), "w") as file:
-            file.write("# Created by joblib automatically.\n")
-            file.write("*\n")
+        gitignore = os.path.join(cache_directory, ".gitignore")
+        if not os.path.exists(gitignore):
+            try:
+                with open(gitignore, "w") as file:
+                    file.write("# Created by joblib automatically.\n")
+                    file.write("*\n")
+            except OSError as e:
+                warnings.warn(f"Unable to write {gitignore}. Exception: {e}.")
 
         # item can be stored compressed for faster I/O
         self.compress = backend_options.get("compress", False)