+++ CHANGES.rst
@@ -4,6 +4,10 @@ Latest changes
 In development
 --------------
 
+- The ``Memory`` object won't overwrite an already existing ``.gitignore`` file in its
+  cache directory anymore.
+  https://github.com/joblib/joblib/pull/1742
+
 - Harden the safety checks in ``eval_expr(pre_dispatch)`` to prevent excessive
   memory allocation and potential crashes by limiting the allowed length of the
   expression and the maximum numeric value of sub-expressions and not